+++ ext/stub/16__closure.zep.c
@@ -17,22 +17,21 @@
 #include "kernel/memory.h"
 
 
-ZEPHIR_INIT_CLASS(stub_11__closure)
+ZEPHIR_INIT_CLASS(stub_16__closure)
 {
-	ZEPHIR_REGISTER_CLASS(stub, 11__closure, stub, 11__closure, stub_11__closure_method_entry, ZEND_ACC_FINAL_CLASS);
+	ZEPHIR_REGISTER_CLASS(stub, 16__closure, stub, 16__closure, stub_16__closure_method_entry, ZEND_ACC_FINAL_CLASS);
 
 	return SUCCESS;
 }
 
-PHP_METHOD(stub_11__closure, __invoke)
+PHP_METHOD(stub_16__closure, __invoke)
 {
 	zephir_method_globals *ZEPHIR_METHOD_GLOBALS_PTR = NULL;
 	zend_long ZEPHIR_LAST_CALL_STATUS;
-	zval *this_ptr = getThis();
 	ZEPHIR_METHOD_GLOBALS_PTR = pecalloc(1, sizeof(zephir_method_globals), 0);
 	zephir_memory_grow_stack(ZEPHIR_METHOD_GLOBALS_PTR, __func__);
 
-	ZEPHIR_RETURN_CALL_METHOD(this_ptr, "issue2497helper", NULL, 0);
+	ZEPHIR_RETURN_CALL_CE_STATIC(stub_mcall_caller_ce, "perform", NULL, 0);
 	zephir_check_call_status();
 	RETURN_MM();
 }